requirements.txt
pandas
numpy
scikit-learn
xgboost
nba_api
tqdm
python-dotenv
requests
